[refactoring] Refactored crawlerFactory.js into multiple files
