+++ lib/factory/index.js
@@ -1,0 +1,1 @@
+exports.createService = require('./service-factory').createService;
