+++ lib/factory/util/logger.js
@@ -1,0 +1,4 @@
+const createLogger = require('../../logging').createLogger;
+const factoryLogger = createLogger(true);
+
+module.exports = factoryLogger;
