+++ lib/server/app.js
@@ -8,13 +8,13 @@ const config = require('painless-config');
 const express = require('express');
 const logger = require('morgan');
 const mockInsights = require('../../providers/logger/mockInsights');
-const CrawlerFactory = require('../crawlerFactory');
+const factory = require('../factory');
 const sendHelper = require('./middleware/sendHelper');
 
 auth.initialize(config.get('CRAWLER_SERVICE_AUTH_TOKEN') || 'secret', config.get('CRAWLER_SERVICE_FORCE_AUTH'));
 mockInsights.setup(config.get('CRAWLER_INSIGHTS_KEY') || 'mock', true);
 const mode = config.get('CRAWLER_MODE') || '';
-const service = CrawlerFactory.createService(mode);
+const service = factory.createService(mode);
 const app = express();
 
 app.disable('x-powered-by');
